Don't store pp if it's the latest for the user in pb

+++ twitterbot/utils.py
@@ -1,5 +1,7 @@
 import os
 from datetime import datetime
+from importlib.metadata import packages_distributions
+from urllib.request import urlopen
 
 import cv2
 import numpy as np
@@ -18,6 +20,70 @@ RAINBOW_BOUNDARIES = [
 ]
 
 USERNAMES = ["exxonmobil", "RogersHelps", "Facebook", "fbsecurity", "SEGA"]
+
+
+def url_to_image(url, readFlag=cv2.IMREAD_COLOR):
+    """
+    Reads an image directly from a URL and returns it
+    """
+
+    # download the image, convert it to a NumPy array, and then read
+    # it into OpenCV format
+    resp = urlopen(url)
+    image = np.asarray(bytearray(resp.read()), dtype="uint8")
+    image = cv2.imdecode(image, readFlag)
+
+    return image
+
+
+def get_current_stored_profile_pic(twitter_username):
+    """
+    For a given Twitter user, get the most recent profile pic
+    stored in pridebot
+    """
+
+    # TODO tromsky: This will likely change once image storage
+    #   is sorted out, but for now this will return the
+    #   the path of the image
+
+    user = TwitterUser.objects.get(username=twitter_username)
+    return (
+        TwitterProfilePic.objects.filter(twitter_user_id=user.id)
+        .order_by("-created_at")[0]
+        .local_path
+    )
+
+
+def get_image(image_path):
+    """
+    Given an image path, return the image as an opencv object
+    """
+
+    return cv2.imread(image_path)
+
+
+def equate_images(first_image, second_image):
+    """
+    Given two image paths, compare the images to see if they are the
+    same. If they are, returns True, else returns False
+    """
+
+    # check size and channels (shape), if these are different, the images are
+    # definitely not the same
+    if first_image.shape != second_image.shape:
+        return False
+
+    # if shapes do match, perform a deeper check
+    # take a difference between the images and split the blue, green and red values
+    # of the difference
+    # if any are not 0, there is a difference
+    difference = cv2.subtract(first_image, second_image)
+    b, g, r = cv2.split(difference)
+    if cv2.countNonZero(b) != 0 or cv2.countNonZero(g) != 0 or cv2.countNonZero(r) != 0:
+        return False
+
+    # at this point, the images match exactly
+    return True
 
 
 def check_image_contains_colours(image_path, colour_boundaries):
@@ -56,6 +122,23 @@ def check_image_contains_colours(image_path, colour_boundaries):
     return image_contains_colours
 
 
+def image_already_latest(username, profile_pic_url):
+    """
+    Returns True if the current profile pic on Twitter matches the most
+    most recent one stored in pridebot
+    """
+
+    # before the image gets stored, check if the image is already stored
+    last_scraped_path = get_current_stored_profile_pic(username)
+
+    last_image = get_image(last_scraped_path)
+    current_image = url_to_image(profile_pic_url)
+
+    images_are_the_same = equate_images(current_image, last_image)
+
+    return images_are_the_same
+
+
 def scrape_profile_pics():
     # build header with bearer token
     bearer_token = os.environ.get("BEARER_TOKEN")
@@ -68,6 +151,11 @@ def scrape_profile_pics():
         profile_pic_path = (
             f"twitterbot/profile_pics/{username}_pp_{datetime.utcnow().isoformat()}.png"
         )
+
+        # if the current profile pic from twitter matches the most recent one
+        # for the user stored in pridebot, don't store it
+        if image_already_latest(username, profile_pic_url):
+            return
 
         # write the profile pic
         with open(profile_pic_path, "wb") as f:
